fix: Use ddm side right
